Fix typos and combine config file for sixgill/shodan
