+++ src/pe_source/__init__.py
@@ -5,4 +5,4 @@
 # Python package.
 from ._version import __version__  # noqa: F401
 
-__all__ = ["cybersixgill"]
+__all__ = ["cybersixgill", "shodan"]
