+++ src/pe_source/_version.py
@@ -1,2 +1,2 @@
 """This file defines the version of this module."""
-__version__ = "1.0.0"
+__version__ = "1.0.1"
