--- src/pe_source/data/shodan/__init__.py
@@ -1,1 +1,0 @@
-"""Initialize the shodan data directory."""
